refactor: handler, util 패키지 -> common 패키지 하위로 이동
handler, util 패키지 -> common 패키지 하위로 이동

- 애플리케이션에서 공통적으로 사용하는 패키지는 common 패키지 하위에서 관리하기 위함
