+++ src/main/java/lotto/common/handler/OutputHandler.java
@@ -1,4 +1,4 @@
-package lotto.handler;
+package lotto.common.handler;
 
 import java.util.stream.Collectors;
 import lotto.model.domain.Lotto;
