+++ src/main/java/lotto/common/util/InputValidator.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.common.util;
 
 import java.util.Objects;
