+++ src/test/java/lotto/InputValidatorTest.java
@@ -2,8 +2,8 @@ package lotto;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import lotto.util.InputParser;
-import lotto.util.InputValidator;
+import lotto.common.util.InputParser;
+import lotto.common.util.InputValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
